backend/app/requirements.txt
gunicorn
torch
torchvision
opencv-python
mtcnn==0.1.1
tensorflow==2.4.1
Flask
tensorflow_addons
Pillow
gfpgan
